Consolidate all env vars into one service `service/env-vars`
