+++ apps/gov-relayer/.env.skel
@@ -1,0 +1,5 @@
+BLOCK_POLLING_INTERVAL=
+CENNZ_NETWORK=
+MESSAGE_MAX_RETRY
+MESSAGE_TIMEOUT=
+PROPOSAL_QUEUE=
