--- apps/gov-website/.env.development
@@ -1,15 +1,0 @@
-DISCORD_BOT_TOKEN=
-DISCORD_CLIENT_ID=987208869770043442
-DISCORD_CLIENT_SECRET=
-DISCORD_IDENTITY_ROLE_ID=
-DISCORD_REGISTRAR_SEED=
-DISCORD_SERVER_ID=
-MONGODB_URI=mongodb://root:root@localhost:27017/admin
-NEXTAUTH_SECRET=
-NEXTAUTH_URL=http://localhost:4200
-NX_CENNZ_NETWORK=rata
-NX_PINATA_GATEWAY=https://gateway.pinata.cloud/ipfs
-NX_PINATA_JWT=
-TWITTER_CLIENT_ID=Sk4zcnBkZk5mVm9OczFPQUU2OXk6MTpjaQ
-TWITTER_CLIENT_SECRET=
-TWITTER_REGISTRAR_SEED=
