+++ apps/gov-website/.env.skel
@@ -1,0 +1,15 @@
+DISCORD_BOT_TOKEN=
+DISCORD_CLIENT_ID=
+DISCORD_CLIENT_SECRET=
+DISCORD_IDENTITY_ROLE_ID=
+DISCORD_REGISTRAR_SEED=
+DISCORD_SERVER_ID=
+MONGODB_URI=
+NEXTAUTH_SECRET=
+NEXTAUTH_URL=
+NX_CENNZ_NETWORK=rata
+NX_PINATA_GATEWAY=
+NX_PINATA_JWT=
+TWITTER_CLIENT_ID=
+TWITTER_CLIENT_SECRET=
+TWITTER_REGISTRAR_SEED=
