--- libs/node/constants/README.md
@@ -1,7 +1,0 @@
-# node-constants
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build node-constants` to build the library.
