--- libs/node/constants/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "@app-gov/node/constants",
-  "version": "0.0.1",
-  "type": "commonjs"
-}
