--- libs/node/constants/project.json
@@ -1,25 +1,0 @@
-{
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/node/constants/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nrwl/js:tsc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/libs/node/constants",
-        "main": "libs/node/constants/src/index.ts",
-        "tsConfig": "libs/node/constants/tsconfig.lib.json",
-        "assets": ["libs/node/constants/*.md"]
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["libs/node/constants/**/*.ts"]
-      }
-    }
-  },
-  "tags": []
-}
