--- libs/service/constants/.babelrc
@@ -1,10 +1,0 @@
-{
-	"presets": [
-		[
-			"@nrwl/web/babel",
-			{
-				"useBuiltIns": "usage"
-			}
-		]
-	]
-}
