--- libs/service/constants/README.md
@@ -1,7 +1,0 @@
-# service-constants
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build service-constants` to build the library.
