+++ libs/service/env-vars/.babelrc
