+++ libs/service/env-vars/.eslintrc.json
