+++ libs/service/env-vars/README.md
@@ -1,0 +1,7 @@
+# service-env-vars
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build service-env-vars` to build the library.
