+++ libs/service/env-vars/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@app-gov/service/constants",
+  "name": "@app-gov/service/env-vars",
   "version": "0.0.1",
   "type": "commonjs"
 }
