+++ libs/service/env-vars/project.json
@@ -1,23 +1,23 @@
 {
   "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/service/constants/src",
+  "sourceRoot": "libs/service/env-vars/src",
   "projectType": "library",
   "targets": {
     "build": {
       "executor": "@nrwl/js:tsc",
       "outputs": ["{options.outputPath}"],
       "options": {
-        "outputPath": "dist/libs/service/constants",
-        "main": "libs/service/constants/src/index.ts",
-        "tsConfig": "libs/service/constants/tsconfig.lib.json",
-        "assets": ["libs/service/constants/*.md"]
+        "outputPath": "dist/libs/service/env-vars",
+        "main": "libs/service/env-vars/src/index.ts",
+        "tsConfig": "libs/service/env-vars/tsconfig.lib.json",
+        "assets": ["libs/service/env-vars/*.md"]
       }
     },
     "lint": {
       "executor": "@nrwl/linter:eslint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["libs/service/constants/**/*.ts"]
+        "lintFilePatterns": ["libs/service/env-vars/**/*.ts"]
       }
     }
   },
