+++ libs/service/env-vars/tsconfig.json
