+++ tsconfig.base.json
@@ -15,11 +15,10 @@
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
     "paths": {
-      "@app-gov/node/constants": ["libs/node/constants/src/index.ts"],
       "@app-gov/node/utils": ["libs/node/utils/src/index.ts"],
       "@app-gov/service/cennznet": ["libs/service/cennznet/src/index.ts"],
-      "@app-gov/service/constants": ["libs/service/constants/src/index.ts"],
       "@app-gov/service/discord": ["libs/service/discord/src/index.ts"],
+      "@app-gov/service/env-vars": ["libs/service/env-vars/src/index.ts"],
       "@app-gov/service/mongodb": ["libs/service/mongodb/src/index.ts"],
       "@app-gov/service/pinata": ["libs/service/pinata/src/index.ts"],
       "@app-gov/service/rabbitmq": ["libs/service/rabbitmq/src/index.ts"],
